:recycle: Update retrofitBearer 코드 수정
Related to : #177
