+++ app/src/main/java/com/umc/anddeul/home/ConfirmDialog.kt
@@ -1,8 +1,6 @@
 package com.umc.anddeul.home
 
 import android.annotation.SuppressLint
-import android.content.Context
-import android.content.SharedPreferences
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
@@ -11,17 +9,15 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
+import com.umc.anddeul.common.RetrofitManager
 import com.umc.anddeul.common.TokenManager
 import com.umc.anddeul.databinding.DialogConfirmBinding
 import com.umc.anddeul.home.model.MemberApproveDTO
 import com.umc.anddeul.home.network.MemberApproveInterface
-import com.umc.anddeul.home.network.MemberInterface
-import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 
 interface ConfirmDialogListener {
     fun onAcceptClicked()
@@ -34,6 +30,7 @@ class ConfirmDialog(name: String, groupName: String, userId: String, private val
     private var groupName : String = ""
     private var userId : String = ""
     var token: String? = null
+    lateinit var retrofitBearer: Retrofit
 
     init {
         this.name = name
@@ -50,6 +47,7 @@ class ConfirmDialog(name: String, groupName: String, userId: String, private val
         binding = DialogConfirmBinding.inflate(layoutInflater, container, false)
 
         token = TokenManager.getToken()
+        retrofitBearer = RetrofitManager.getRetrofitInstance()
 
         dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT)) // 배경 투명
         binding.dialogNameTv.text = "'${name}'님을"
@@ -69,22 +67,6 @@ class ConfirmDialog(name: String, groupName: String, userId: String, private val
     }
 
     fun approveMember(userId: String) {
-        val retrofitBearer = Retrofit.Builder()
-            .baseUrl("http://umc-garden.store")
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(
-                OkHttpClient.Builder()
-                    .addInterceptor { chain ->
-                        val request = chain.request().newBuilder()
-                            .addHeader("Authorization", "Bearer " + token.orEmpty())
-                            .build()
-                        Log.d("retrofitBearer", "Token: ${token.toString()}" + token.orEmpty())
-                        chain.proceed(request)
-                    }
-                    .build()
-            )
-            .build()
-
         val approveService = retrofitBearer.create(MemberApproveInterface::class.java)
 
         approveService.approveMember(userId).enqueue(object : Callback<MemberApproveDTO> {
@@ -104,9 +86,6 @@ class ConfirmDialog(name: String, groupName: String, userId: String, private val
             override fun onFailure(call: Call<MemberApproveDTO>, t: Throwable) {
                 Log.e("approveService", "Failure message: ${t.message}")
             }
-
         })
-
-
     }
 }
