+++ app/src/main/java/com/umc/anddeul/home/DeleteDialog.kt
@@ -1,7 +1,5 @@
 package com.umc.anddeul.home
 
-import android.content.Context
-import android.content.SharedPreferences
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
@@ -10,20 +8,20 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
+import com.umc.anddeul.common.RetrofitManager
 import com.umc.anddeul.common.TokenManager
 import com.umc.anddeul.databinding.FragmentDialogPermissionBinding
 import com.umc.anddeul.home.model.PostDelete
 import com.umc.anddeul.home.network.PostDeleteInterface
-import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 
 class DeleteDialog(val postId : Int) : DialogFragment() {
     lateinit var binding: FragmentDialogPermissionBinding
     var token : String? = null
+    lateinit var retrofitBearer: Retrofit
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -33,6 +31,7 @@ class DeleteDialog(val postId : Int) : DialogFragment() {
         binding = FragmentDialogPermissionBinding.inflate(layoutInflater, container, false)
 
         token = TokenManager.getToken()
+        retrofitBearer = RetrofitManager.getRetrofitInstance()
 
         dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT)) // 배경 투명
         binding.dialogPermissionTv.text = "게시물을 삭제 하시겠어요?"
@@ -51,23 +50,6 @@ class DeleteDialog(val postId : Int) : DialogFragment() {
     }
 
     fun delete(postId : Int) {
-
-        val retrofitBearer = Retrofit.Builder()
-            .baseUrl("http://umc-garden.store")
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(
-                OkHttpClient.Builder()
-                    .addInterceptor { chain ->
-                        val request = chain.request().newBuilder()
-                            .addHeader("Authorization", "Bearer " + token.orEmpty())
-                            .build()
-                        Log.d("retrofitBearer", "Token: ${token.toString()}" + token.orEmpty())
-                        chain.proceed(request)
-                    }
-                    .build()
-            )
-            .build()
-
         val deleteService = retrofitBearer.create(PostDeleteInterface::class.java)
 
         deleteService.deletePost(postId).enqueue(object : Callback<PostDelete> {
@@ -86,13 +68,11 @@ class DeleteDialog(val postId : Int) : DialogFragment() {
                         homeFragment?.loadPost()
                     }
                 }
-
             }
 
             override fun onFailure(call: Call<PostDelete>, t: Throwable) {
                 Log.e("deleteService", "Failure message: ${t.message}")
             }
-
         })
     }
 }
