+++ app/src/main/java/com/umc/anddeul/home/PostRVAdapter.kt
@@ -1,7 +1,6 @@
 package com.umc.anddeul.home
 
 import android.content.Context
-import android.content.SharedPreferences
 import android.util.Log
 import android.view.Gravity
 import android.view.LayoutInflater
@@ -12,24 +11,21 @@ import android.widget.PopupMenu
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.widget.ViewPager2
 import com.umc.anddeul.R
+import com.umc.anddeul.common.RetrofitManager
 import com.umc.anddeul.common.TokenManager
 import com.umc.anddeul.databinding.FragmentHomeMyUploadBinding
 import com.umc.anddeul.databinding.FragmentHomeUploadBinding
 import com.umc.anddeul.home.model.EmojiDTO
 import com.umc.anddeul.home.model.EmojiRequest
-import com.umc.anddeul.home.model.EmojiResult
 import com.umc.anddeul.home.model.PostData
 import com.umc.anddeul.home.network.EmojiInterface
-import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.create
 
 class PostRVAdapter(private val context: Context, var postList: List<PostData>, var authorTypeList: List<String>) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
     var token = TokenManager.getToken()
+    var retrofitBearer = RetrofitManager.getRetrofitInstance()
 
     companion object {
         const val VIEW_TYPE_ME = 0
@@ -231,21 +227,6 @@ class PostRVAdapter(private val context: Context, var postList: List<PostData>, 
         // 사라지는 애니메이션
         val fadeOutAnimation = AnimationUtils.loadAnimation(context, R.anim.fade_out)
 
-        val retrofitBearer = Retrofit.Builder()
-            .baseUrl("http://umc-garden.store")
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(
-                OkHttpClient.Builder()
-                    .addInterceptor { chain ->
-                        val request = chain.request().newBuilder()
-                            .addHeader("Authorization", "Bearer " + token.orEmpty())
-                            .build()
-                        chain.proceed(request)
-                    }
-                    .build()
-            )
-            .build()
-
         val emojiService = retrofitBearer.create(EmojiInterface::class.java)
         val emojiRequest = EmojiRequest(emojiType)
 
@@ -308,21 +289,6 @@ class PostRVAdapter(private val context: Context, var postList: List<PostData>, 
     fun selectMyEmoji(binding: FragmentHomeMyUploadBinding, postId: Int, emojiType: String) {
         // 사라지는 애니메이션
         val fadeOutAnimation = AnimationUtils.loadAnimation(context, R.anim.fade_out)
-
-        val retrofitBearer = Retrofit.Builder()
-            .baseUrl("http://umc-garden.store")
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(
-                OkHttpClient.Builder()
-                    .addInterceptor { chain ->
-                        val request = chain.request().newBuilder()
-                            .addHeader("Authorization", "Bearer " + token.orEmpty())
-                            .build()
-                        chain.proceed(request)
-                    }
-                    .build()
-            )
-            .build()
 
         val emojiService = retrofitBearer.create(EmojiInterface::class.java)
         val emojiRequest = EmojiRequest(emojiType)
