+++ app/src/main/java/com/umc/anddeul/home/UserProfileRVAdapter.kt
@@ -1,6 +1,5 @@
 package com.umc.anddeul.home
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
