+++ app/src/main/java/com/umc/anddeul/mypage/LogoutDialog.kt
@@ -1,8 +1,6 @@
 package com.umc.anddeul.mypage
 
-import android.content.Context
 import android.content.Intent
-import android.content.SharedPreferences
 import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import android.os.Bundle
@@ -11,21 +9,21 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.DialogFragment
+import com.umc.anddeul.common.RetrofitManager
 import com.umc.anddeul.common.TokenManager
 import com.umc.anddeul.databinding.FragmentDialogPermissionBinding
 import com.umc.anddeul.mypage.model.LogoutDTO
 import com.umc.anddeul.mypage.network.LogoutInterface
 import com.umc.anddeul.start.StartActivity
-import okhttp3.OkHttpClient
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
 
 class LogoutDialog : DialogFragment() {
     lateinit var binding: FragmentDialogPermissionBinding
     var token: String? = null
+    lateinit var retrofitBearer: Retrofit
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -35,6 +33,7 @@ class LogoutDialog : DialogFragment() {
         binding = FragmentDialogPermissionBinding.inflate(layoutInflater, container, false)
 
         token = TokenManager.getToken()
+        retrofitBearer = RetrofitManager.getRetrofitInstance()
 
         dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT)) // 배경 투명
         binding.dialogPermissionTv.text = "로그아웃 하시겠어요?"
@@ -54,22 +53,6 @@ class LogoutDialog : DialogFragment() {
     }
 
     fun myPageLogout() {
-        val retrofitBearer = Retrofit.Builder()
-            .baseUrl("http://umc-garden.store")
-            .addConverterFactory(GsonConverterFactory.create())
-            .client(
-                OkHttpClient.Builder()
-                    .addInterceptor { chain ->
-                        val request = chain.request().newBuilder()
-                            .addHeader("Authorization", "Bearer " + token.orEmpty())
-                            .build()
-                        Log.d("retrofitBearer", "Token: ${token.toString()}" + token.orEmpty())
-                        chain.proceed(request)
-                    }
-                    .build()
-            )
-            .build()
-
         val logoutService = retrofitBearer.create(LogoutInterface::class.java)
 
         logoutService.logout().enqueue(object : Callback<LogoutDTO> {
